# Suggestion: smarter witness calculation