There was a suggestion to use a `Map Natural (Map Natural a -> a)` type for witness generation. This PR implements this suggestion; evaluation of a final witness is done via the trick described in https://github.com/quchen/articles/blob/master/loeb-moeb.md.
While this works, I would like to benchmark this solution against the current implementation somehow.