@echatav 2024-07-11
Maybe timing/running (some subset of) tests could provide a benchmark.