@TurtlePU 2024-07-12
Running the `arithmetic-circuit` benchmark for SHA2 hashing with 2 rounds yields:

Before:
```
  Calculating SHA2 512/364 of a bytestring of length 32:  OK
    63.3 ms ± 2.1 ms
  Calculating SHA2 512/364 of a bytestring of length 64:  OK
    58.5 ms ± 1.8 ms
  Calculating SHA2 512/364 of a bytestring of length 128: OK
    59.1 ms ± 4.0 ms
  Calculating SHA2 512/364 of a bytestring of length 256: OK
    57.9 ms ± 3.5 ms
  Calculating SHA2 512/364 of a bytestring of length 512: OK
    4.597 s ± 343 ms

All 5 tests passed (33.21s)
```

After:
```
  Calculating SHA2 512/364 of a bytestring of length 32:  OK
    4.18 ms ± 239 μs
  Calculating SHA2 512/364 of a bytestring of length 64:  OK
    4.04 ms ± 338 μs
  Calculating SHA2 512/364 of a bytestring of length 128: OK
    4.05 ms ± 387 μs
  Calculating SHA2 512/364 of a bytestring of length 256: OK
    4.10 ms ± 193 μs
  Calculating SHA2 512/364 of a bytestring of length 512: OK
    10.5 ms ± 799 μs

All 5 tests passed (18.59s)
```