@TurtlePU 2024-07-12
Honest SHA2 with 64 rounds benchmarks fine:
```
  Calculating SHA2 512/364 of a bytestring of length 32:  OK
    255  ms ± 3.4 ms
  Calculating SHA2 512/364 of a bytestring of length 64:  OK
    260  ms ±  11 ms
  Calculating SHA2 512/364 of a bytestring of length 128: OK
    266  ms ± 5.4 ms
  Calculating SHA2 512/364 of a bytestring of length 256: OK
    276  ms ±  26 ms
  Calculating SHA2 512/364 of a bytestring of length 512: OK
    678  ms ±  60 ms

All 5 tests passed (1487.99s)
```